Sensor Polling
Implemented the polling on the UartSensor class. Off course this should
eventually be in the ISensor class so all sensors can profit from it,
but I didn't need it, so didn't implement it. One can now subscribe to a
sensor's event and the queuethread will decide when an event gets
raised.